Move Type, TypeBound, PrimType to better locations (#403)
namely Type, TypeBound -> types.rs
PrimType -> primitive.rs

also add docstrings
